Simulated implementation of JerryScript API using Emscripten
Add a re-implementation of the JerryScript API in Emscripten ASM and JavaScript. This makes it possible to "transpile" C code that uses the JerryScript API into .js code that can be run inside any JavaScript environment, like a web browser, without having to transpile the entire JerryScript C code base as well. The host JavaScript environment is used to perform any JerryScript API call that the C code may be using.

JerryScript-DCO-1.0-Signed-off-by: Martijn The martijn.the@intel.com